requirements.txt
Django
djangorestframework
django-filter
redis
celery
mem0ai
gunicorn
python-dotenv
psycopg2-binary
gunicorn